+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/Protocol/Rdp/TestSessionContextFactoryForRdp.cs
@@ -102,7 +102,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Protocol.Rdp
                 new Mock<IAuthorization>().Object,
                 workspace.Object,
                 new Mock<IKeyStoreAdapter>().Object,
-                new Mock<IKeyAuthorizationService>().Object,
+                new Mock<IKeyAuthorizer>().Object,
                 settingsService.Object,
                 new Mock<IIapTransportFactory>().Object,
                 new Mock<IDirectTransportFactory>().Object,
@@ -153,7 +153,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Protocol.Rdp
                 new Mock<IAuthorization>().Object,
                 workspace.Object,
                 new Mock<IKeyStoreAdapter>().Object,
-                new Mock<IKeyAuthorizationService>().Object,
+                new Mock<IKeyAuthorizer>().Object,
                 settingsService.Object,
                 new Mock<IIapTransportFactory>().Object,
                 new Mock<IDirectTransportFactory>().Object,
@@ -209,7 +209,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Protocol.Rdp
                 new Mock<IAuthorization>().Object,
                 workspace.Object,
                 new Mock<IKeyStoreAdapter>().Object,
-                new Mock<IKeyAuthorizationService>().Object,
+                new Mock<IKeyAuthorizer>().Object,
                 settingsService.Object,
                 new Mock<IIapTransportFactory>().Object,
                 new Mock<IDirectTransportFactory>().Object,
@@ -263,7 +263,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Protocol.Rdp
                 new Mock<IAuthorization>().Object,
                 workspace.Object,
                 new Mock<IKeyStoreAdapter>().Object,
-                new Mock<IKeyAuthorizationService>().Object,
+                new Mock<IKeyAuthorizer>().Object,
                 settingsService.Object,
                 new Mock<IIapTransportFactory>().Object,
                 new Mock<IDirectTransportFactory>().Object,
@@ -328,7 +328,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Protocol.Rdp
                 new Mock<IAuthorization>().Object,
                 workspace.Object,
                 new Mock<IKeyStoreAdapter>().Object,
-                new Mock<IKeyAuthorizationService>().Object,
+                new Mock<IKeyAuthorizer>().Object,
                 settingsService.Object,
                 new Mock<IIapTransportFactory>().Object,
                 new Mock<IDirectTransportFactory>().Object,
@@ -375,7 +375,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Protocol.Rdp
                 new Mock<IAuthorization>().Object,
                 new Mock<IProjectWorkspace>().Object,
                 new Mock<IKeyStoreAdapter>().Object,
-                new Mock<IKeyAuthorizationService>().Object,
+                new Mock<IKeyAuthorizer>().Object,
                 new Mock<IConnectionSettingsService>().Object,
                 new Mock<IIapTransportFactory>().Object,
                 new Mock<IDirectTransportFactory>().Object,
@@ -413,7 +413,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Protocol.Rdp
                 new Mock<IAuthorization>().Object,
                 new Mock<IProjectWorkspace>().Object,
                 new Mock<IKeyStoreAdapter>().Object,
-                new Mock<IKeyAuthorizationService>().Object,
+                new Mock<IKeyAuthorizer>().Object,
                 new Mock<IConnectionSettingsService>().Object,
                 new Mock<IIapTransportFactory>().Object,
                 new Mock<IDirectTransportFactory>().Object,
