+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/Protocol/Ssh/TestSessionContextFactoryForSsh.cs
@@ -130,7 +130,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Protocol.Ssh
                 CreateAuthorizationMock().Object,
                 CreateProjectModelServiceMock(OperatingSystems.Linux).Object,
                 keyStore.Object,
-                new Mock<IKeyAuthorizationService>().Object,
+                new Mock<IKeyAuthorizer>().Object,
                 settingsService.Object,
                 new Mock<IIapTransportFactory>().Object,
                 new Mock<IDirectTransportFactory>().Object,
@@ -175,7 +175,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Protocol.Ssh
                 CreateAuthorizationMock().Object,
                 CreateProjectModelServiceMock(OperatingSystems.Linux).Object,
                 CreateKeyStoreAdapterMock().Object,
-                new Mock<IKeyAuthorizationService>().Object,
+                new Mock<IKeyAuthorizer>().Object,
                 settingsService.Object,
                 new Mock<IIapTransportFactory>().Object,
                 new Mock<IDirectTransportFactory>().Object,
@@ -221,7 +221,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Protocol.Ssh
                 CreateAuthorizationMock().Object,
                 CreateProjectModelServiceMock(OperatingSystems.Linux).Object,
                 CreateKeyStoreAdapterMock().Object,
-                new Mock<IKeyAuthorizationService>().Object,
+                new Mock<IKeyAuthorizer>().Object,
                 settingsService.Object,
                 new Mock<IIapTransportFactory>().Object,
                 new Mock<IDirectTransportFactory>().Object,
