+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/Views/SshTerminal/TestSshTerminalPaneViewModel.cs
@@ -105,7 +105,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Test.Views.SshTerminal
         {
             var authorizationSource = CreateAuthorizationMock();
 
-            var keyAdapter = new KeyAuthorizationService(
+            var keyAdapter = new KeyAuthorizer(
                 authorizationSource.Object,
                 new ComputeEngineAdapter(credential.ToAuthorization(), TestProject.UserAgent),
                 new ResourceManagerAdapter(credential.ToAuthorization(), TestProject.UserAgent),
