+++ sources/Google.Solutions.IapDesktop.Extensions.Session/Google.Solutions.IapDesktop.Extensions.Session.csproj
@@ -159,8 +159,7 @@
     <Compile Include="Settings\SshSettings.cs" />
     <Compile Include="Settings\TerminalSettings.cs" />
     <Compile Include="Protocol\Ssh\AuthorizedKeyPair.cs" />
-    <Compile Include="Protocol\Ssh\KeyAuthorizationService.cs" />
-    <Compile Include="Protocol\Ssh\IKeyAuthorizationService.cs" />
+    <Compile Include="Protocol\Ssh\KeyAuthorizer.cs" />
     <Compile Include="Protocol\Ssh\MetadataAuthorizedPublicKeyProcessor.cs" />
     <Compile Include="Protocol\Ssh\OsLoginProfile.cs" />
     <Compile Include="Settings\ConnectionSettingsService.cs" />
